stop bindings some types & functions to python
The types and functions from

  almath/dsp/digitalfilter.h
  almath/dsp/pidcontroller.h
  almath/types/occupancymapparams.h

Were wrapped in python, but those bindings are not
tested, nor used by any python documentation example.
Those types are also not used by any documented service API.

They were most probably added to the bindings for internal
prototyping and debugging.
As such, they might belong to the almathinternal python bindings,
but not to the almath ones.

Change-Id: I2e43673a9bca8f804c4946aea88f1f11dc2b45e4